+++ HISTORY.md
@@ -1,4 +1,15 @@
 # History
 
+## 0.3.0 (2019-07-11)
+* Add Source Group search parameter
+* Add Satellite Name search parameter
+
+## 0.2.1 (2019-07-03)
+* Documentation updates
+* download done queue added to bulk downloader
+
+## 0.2.0 (2019-07-02)
+* Initial release on PyPI
+
 ## 0.0.1 (2019-06-04)
 * Initial code release without packaging.